requirements.txt
alembic==1.12.0
asyncio==3.4.3
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Email==1.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Reuploaded==1.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
Pillow==9.5.0
PyJWT==2.8.0
python-dotenv==1.0.0
sentry-sdk==1.31.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.0.5
Werkzeug==2.3.7
WTForms==3.0.1
